Move spinner to auth directory
