+++ src/components/Auth/Spinner.scss
